Add task to create log file
Add a new task to create a log file using the `file` module in Ansible. The task creates a file at the path `/var/log/simplesaml/simplesamlphp.log` with the mode `0664`, owned by the `apache` user and group. This will ensure that the log file is created with the correct permissions and ownership for Apache to write to.