Refactor simplesaml config file to use templates
This commit moves the config.php file for the simplesaml role to a new templates directory. This change allows for the file to be more easily managed and templated for different environments.

This commit also  updates the simplesaml role to use Ansible to copy the config.php file from the roles/simplesaml/templates/ directory to its final destination. This change ensures that the file is properly managed and templated in a repeatable and automated way.