Add HTTPS VHost to xdmod.conf with updated variables in group_vars/all
This commit includes changes to the xdmod.conf file within the openxdmod role, which adds a virtual host (VHost) for HTTPS. The necessary variables have been updated in group_vars/all to support this change.

Specifically, the changes to the xdmod.conf file include the addition of a new server block with the appropriate SSL configurations for HTTPS. In group_vars/all, variables such as SSL certificates and keys have been updated to reflect this new configuration.