Add Apache HTTPS virtual host configuration
This commit adds an Apache virtual host configuration file for serving XDMod over HTTPS. The configuration file sets up a permanent redirect from HTTP to HTTPS, and includes SSL certificate and key file paths. Additionally, it specifies the document root, directory options, and access control for the application, along with an alias for SimpleSAMLphp. Error and access logging is also configured for the virtual host. This file is intended to be used in conjunction with the existing XDMod Apache configuration.

This commit also adds this template as a part of the openssl_cert_gen role to put the modified xdmod apache file in the system location /etc/httpd/conf.d/