Add [authentication] section to portal_settings.ini
This commit adds a new task to the `simplesaml` role to insert a `[authentication]` section to `portal_settings.ini` configuration file. The new task uses the `blockinfile` module to add the new section on a new line after the `basic_auth = "on"` line. The `marker` parameter is used to add a comment to the file to indicate that the block was added by Ansible and should not be edited manually.