Add flag for SAML and UAB IDP selection and update role to use UAB IDP from S3 bucket
In this commit:

Added a flag in groupvars/all to switch between SAML and UAB IDP
Updated the role to utilize UAB IDP
UAB IDP configuration files are now stored in an S3 bucket
This commit enhances the system's flexibility by allowing the selection of either SAML or UAB IDP based on the flag in groupvars/all. The role has been modified to utilize the UAB IDP, which is now conveniently stored in an S3 bucket for easy access and management.