Refactor: Update user_manual link in configuration
In this commit, the `user_manual` configuration parameter has been updated to point to the new user manual location. Previously, it was set as `{{ xdmod_address }} user_manual`, and now it points to `{{ xdmod_address }}user_manual/index.php`. This change ensures that the user manual link correctly directs users to the updated location.